requirements.txt
pandas==2.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
ratelimit==2.2.1
tqdm==4.66.1
countrynames==1.13.0
unidecode==1.3.6
streamlit==1.32.0
plotly==5.18.0
seaborn==0.13.2
matplotlib==3.8.3 